+++ src/main/java/com/zerobase/hseungho/restaurantreservation/service/controller/UserController.java
@@ -1,8 +1,8 @@
 package com.zerobase.hseungho.restaurantreservation.service.controller;
 
 import com.zerobase.hseungho.restaurantreservation.service.appservice.UserService;
-import com.zerobase.hseungho.restaurantreservation.service.dto.external.user.CheckUsingResourceAvailable;
-import com.zerobase.hseungho.restaurantreservation.service.dto.external.user.SignUp;
+import com.zerobase.hseungho.restaurantreservation.service.dto.CheckUsingResourceAvailable;
+import com.zerobase.hseungho.restaurantreservation.service.dto.SignUp;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
