+++ src/main/java/com/zerobase/hseungho/restaurantreservation/service/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.zerobase.hseungho.restaurantreservation.service.dto.internal;
+package com.zerobase.hseungho.restaurantreservation.service.dto;
 
 import com.zerobase.hseungho.restaurantreservation.service.domain.User;
 import com.zerobase.hseungho.restaurantreservation.service.type.UserType;
